bi_lms/requirements.txt
django-admin-interface
django-ckeditor
edx-organizations
# mysqlclient a través de pip install mysqlcient
# mysqlclient a través de pip install djangorestframework
